+++ examples/terminal.rs
@@ -11,7 +11,7 @@ struct Term {
 }
 
 impl Term {
-    pub fn new(buf: &mut TextBuffer) -> Term {
+    pub fn new(buf: TextBuffer) -> Term {
         let mut current_dir = std::env::current_dir()
             .unwrap()
             .to_string_lossy()
@@ -105,9 +105,9 @@ impl DerefMut for Term {
 fn main() {
     let app = app::App::default().set_scheme(app::AppScheme::Plastic);
     let mut wind = Window::new(100, 100, 640, 480, "Rusty Terminal");
-    let mut buf = TextBuffer::default();
+    let buf = TextBuffer::default();
 
-    let mut term = Term::new(&mut buf);
+    let mut term = Term::new(buf);
     term.style();
 
     let dir = term.current_dir.clone();