+++ fltk-sys/cfltk/cfl_menu.h
@@ -93,6 +93,10 @@ void Fl_Menu_Item_deactivate(Fl_Menu_Item *);
 
 int Fl_Menu_Item_submenu(const Fl_Menu_Item *self);
 
+int Fl_Menu_Item_checkbox(const Fl_Menu_Item *self);
+
+int Fl_Menu_Item_radio(const Fl_Menu_Item *self);
+
 Fl_Menu_Item *Fl_Menu_Item_next(Fl_Menu_Item *self, int idx);
 
 void Fl_Menu_Item_callback(Fl_Menu_Item *self, Fl_Callback *c, void *p);