+++ fltk-sys/cfltk/cfl_window.h
@@ -16,7 +16,9 @@ extern "C" {
     void widget##_make_resizable(widget *self, void *);                                            \
     void widget##_set_cursor(widget *self, int cursor);                                            \
     int widget##_shown(widget *self);                                                              \
-    void *widget##_raw_handle(const widget *w);
+    void *widget##_raw_handle(const widget *w);                                                    \
+    void widget##_set_border(widget *, int flag);\
+    int widget##_border(const widget *);
 
 WIDGET_DECLARE(Fl_Window)